Clear previous element info.
